Refactor User.js into two different files
