+++ api/mailer/mailer.js
@@ -2,7 +2,6 @@ const express = require('express');
 const router = express.Router();
 const { SceGoogleApiHandler } = require('./SceGoogleApiHandler');
 const verification = require('./templates/verification');
-const { validateVerificationEmail, generateHashedId } = require('./auth');
 const {
   OK,
   BAD_REQUEST
@@ -38,17 +37,6 @@ router.post('/sendVerificationEmail', async (req, res) => {
         });
     })
     .catch(_ => {
-      res.sendStatus(BAD_REQUEST);
-    });
-});
-
-router.post('/validateVerificationEmail', async (req, res) => {
-  validateVerificationEmail(
-    req.body.email, req.body.hashedId)
-    .then((_) => {
-      res.sendStatus(OK);
-    })
-    .catch((err) => {
       res.sendStatus(BAD_REQUEST);
     });
 });
