+++ src/Pages/Login/Login.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { Row, Container } from 'reactstrap';
 import { Link } from 'react-router-dom';
 import LoginInput from './LoginInput';
-import { loginUser } from '../../APIFunctions/User';
+import { loginUser } from '../../APIFunctions/Auth';
 import './login.css';
 
 export default function Login(props) {
